requirements.txt
alembic==1.12.1
bcrypt==4
email-validator==2.0.0
fastapi==0.104.1
httpx==0.25.1
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic-settings==2.0.3
pydantic==2.4.2
pytest-cov==4.1.0
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
requests==2.31.0
sqlalchemy==2.0.23
structlog==23.2.0
uvicorn==0.23.2